Finished ClientHelloMessage and ServerHelloMessage creation

+++ src/net/ClientHelloMessage.h
@@ -2,12 +2,13 @@
 #include "net/AbstractMessage.h"
 
 namespace net {
+	
 	class ClientHelloMessage : public AbstractMessage {
 	public:
-		ClientHelloMessage(unsigned int port);
+		ClientHelloMessage(unsigned short port);
 		void createBuffer(struct Message* msg);
 
 	private:
-		unsigned int port;
+		unsigned short port;
 	};
 }
